refactor: RequestLineParserTest, RequestLineTest 이동
